Protocol version 32 (12w18a)
This commit implements the RC4 encryption added in 12w17a as well as the AES encryption which will be used in future versions. Not supported is the HC-256 cipher used in 12w19a. The new encryption requires a proxy to authenticate itself with a valid minecraft account on servers in online-mode. This can be achieved with one of the three new options: --auto-authenticate, --user, --password-file.