network.Server: use handler.run() instead of handler.start()
According to http://www.gevent.org/gevent.baseserver.html, when the handler 
returns the socket is closed. `Greenlet.start()` is a scheduling function and
returns immediately, which in turn closes the socket. As Minecraft use a
persistent connection, closing the socket would be undesireable. Instead, 
`Greenlet.run()` calls `Endpoint._run()`, which waits for the handler to
finish execution, keeping the connection persistent.